Add support for managing more configuration files
Beyond the main opensearch.yml configuration file, OpenSearch has a
bunch of files for the many plugins it ships with.  Add support for
managing these configuration files a similar way it is possible to
manage opensearch.yml: ships default config in Hiera, allow the user to
pass their own config and allow them to choose if the user supplied
config should be merged into the default or replace them completely.

Some configuration changes are noop after the initial bootstrap, but
this is considered out-of-scope for now.

In order to avoid drift between the default configuration shipped with
the module and the actual default configuration of the upstream package,
add a GitHub action that open PR when the defaults need updating.