Fix GitHub action
This repository has a GitHub action to maintain the default
configuration of OpenSearch in sync with the upstream project.

Upstream has removed its 'whitelist' settings in favor of the
`allowlist` ones, which result in failure of the action to gather the
default configuration:
https://github.com/opensearch-project/security/pull/5224

Adjust the update script so that the removed configuration is not looked
for.